pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "modelc"
version = "0.1.0"
description = "CLI for portable, inspectable AI model containers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.3",
  "PyYAML>=6.0.1",
  "pydantic>=2.7.0"
]

[project.scripts]
modelc = "modelc.cli:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["modelc*"]
